+++ test/dantzig/dsl/edge_case_tests.exs
@@ -1,0 +1,172 @@
+defmodule Dantzig.DSL.EdgeCaseTests do
+  use ExUnit.Case
+
+  alias Dantzig.Problem
+
+  describe "Edge Cases and Error Conditions" do
+    test "empty generators should raise error" do
+      assert_raise ArgumentError, fn ->
+        Problem.define do
+          new(direction: :maximize)
+          variables("x", [], :continuous)  # Empty generator list
+        end
+      end
+    end
+
+    test "invalid variable type should raise error" do
+      assert_raise ArgumentError, fn ->
+        Problem.define do
+          new(direction: :maximize)
+          variables("x", [i <- 1..3], :invalid_type)
+        end
+      end
+    end
+
+    test "invalid constraint operator should raise error" do
+      assert_raise ArgumentError, fn ->
+        Problem.define do
+          new(direction: :maximize)
+          variables("x", :continuous)
+          constraints(x < 5)  # Invalid operator
+        end
+      end
+    end
+
+    test "undefined variable in constraint should raise error" do
+      assert_raise ArgumentError, fn ->
+        Problem.define do
+          new(direction: :maximize)
+          variables("x", :continuous)
+          constraints(y <= 5)  # y is not defined
+        end
+      end
+    end
+
+    test "invalid objective direction should raise error" do
+      assert_raise ArgumentError, fn ->
+        Problem.define do
+          new(direction: :invalid)
+          variables("x", :continuous)
+          objective(x)
+        end
+      end
+    end
+
+    test "very large problem should handle gracefully" do
+      # Test with larger but reasonable problem size
+      problem =
+        Problem.define do
+          new(direction: :maximize)
+          variables("x", [i <- 1..20], :binary)
+          constraints([i <- 1..20], x(i) <= 1)
+          objective(sum(for i <- 1..20, do: x(i)))
+        end
+
+      # Should not crash during problem creation
+      assert problem != nil
+      assert map_size(problem.variables) > 0
+    end
+
+    test "constraint with no variables should handle gracefully" do
+      problem =
+        Problem.define do
+          new(direction: :maximize)
+          variables("x", :continuous)
+          constraints(5 <= 10)  # Constant constraint
+          objective(x)
+        end
+
+      assert problem != nil
+    end
+
+    test "problem with no constraints should work" do
+      problem =
+        Problem.define do
+          new(direction: :maximize)
+          variables("x", :continuous)
+          objective(x)
+        end
+
+      assert problem != nil
+      assert problem.direction == :maximize
+    end
+  end
+
+  describe "Boundary Conditions" do
+    test "single variable problem" do
+      problem =
+        Problem.define do
+          new(direction: :maximize)
+          variables("x", :continuous, min: 0, max: 10)
+          constraints(x <= 5)
+          objective(x)
+        end
+
+      assert problem != nil
+      assert map_size(problem.variables) == 1
+    end
+
+    test "zero bounds" do
+      problem =
+        Problem.define do
+          new(direction: :maximize)
+          variables("x", :continuous, min: 0, max: 0)
+          constraints(x == 0)
+          objective(x)
+        end
+
+      assert problem != nil
+    end
+
+    test "very small coefficients" do
+      problem =
+        Problem.define do
+          new(direction: :maximize)
+          variables("x", :continuous)
+          variables("y", :continuous)
+          constraints(0.0001*x + 0.0001*y <= 1)
+          objective(x + y)
+        end
+
+      assert problem != nil
+    end
+  end
+
+  describe "Complex Expression Handling" do
+    test "nested expressions" do
+      problem =
+        Problem.define do
+          new(direction: :maximize)
+          variables("x", :continuous)
+          variables("y", :continuous)
+          variables("z", :continuous)
+
+          # Complex nested constraint
+          constraints(2*(x + y) - 3*z <= 10)
+          objective(x + 2*y + 3*z)
+        end
+
+      assert problem != nil
+    end
+
+    test "expression with many terms" do
+      problem =
+        Problem.define do
+          new(direction: :maximize)
+          variables("x1", :continuous)
+          variables("x2", :continuous)
+          variables("x3", :continuous)
+          variables("x4", :continuous)
+          variables("x5", :continuous)
+
+          # Constraint with many variables
+          constraints(x1 + x2 + x3 + x4 + x5 <= 100)
+          objective(x1 + x2 + x3 + x4 + x5)
+        end
+
+      assert problem != nil
+    end
+  end
+end    end
+  end
+end
